+++ src/pages/ListEditora.js
@@ -1,9 +1,10 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 import {Alert} from 'react-native'
 import {
     Container, Header, Title, Content, Button, Left,
     Body, Icon, Text, Card, CardItem, View, Footer, FooterTab,
 } from 'native-base';
+import ActionButtonCustom from '../components/ActionButtonCustom';
 import { FlatList } from 'react-native-gesture-handler';
 import api from '../services/api';
 
@@ -13,7 +14,9 @@ export default function ListEditora(props) {
         const response = await api.get('/editoras');
         setEditoras(response.data);
     }
-    carregarEditoras();
+    useEffect(()=>{
+        carregarEditoras()
+    },[])
     return (
         <Container>
             <View style={{ backgroundColor: '#1A237E', height: 23 }}></View>
@@ -40,8 +43,19 @@ export default function ListEditora(props) {
                                         </CardItem>
                                         <CardItem footer>
                                             <Button danger onPress={async () => {
-                                                const id = item.id;
-                                                await api.delete(`/editoras/${id}`);
+                                                Alert.alert(
+                                                    'Excluir',
+                                                    'Confimar a exclusão ?',
+                                                    [
+                                                        { text: 'Não', style: 'cancel' },
+                                                        {
+                                                            text: 'Sim', onPress: async () => {
+                                                                const id = item.id;
+                                                                await api.delete(`/editoras/${id}`);
+                                                            }
+                                                        }
+                                                    ]
+                                                )
                                             }} >
                                                 <Text>Excluir</Text>
                                             </Button>
@@ -52,13 +66,7 @@ export default function ListEditora(props) {
                         )} />
                 </View>
             </Content>
-            <Footer>
-                <FooterTab>
-                    <Button full>
-                        <Text>Footer</Text>
-                    </Button>
-                </FooterTab>
-            </Footer>
+            <ActionButtonCustom props={props} title='Novo' route='FormEditora'/>
         </Container>
     );
 }
